Revert "db intent"
This reverts commit cde13ab4eaeb9eb1563e316d644fb1834acb6516.
