+++ app/src/main/java/com/example/apparty/CompletePurchaseFragment.java
@@ -1,4 +1,4 @@
-package com.example.apparty.fragments;
+package com.example.apparty;
 
 import android.os.Bundle;
 
@@ -8,14 +8,11 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.apparty.databinding.FragmentDetailEventItemBinding;
+public class CompletePurchaseFragment extends Fragment {
 
-public class DetailEventItemFragment extends Fragment {
-
-    private FragmentDetailEventItemBinding binding;
-
-    public DetailEventItemFragment() {}
-
+    public CompletePurchaseFragment() {
+        // Required empty public constructor
+    }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -25,8 +22,7 @@ public class DetailEventItemFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        binding = FragmentDetailEventItemBinding.inflate(inflater, container, false);
-        return binding.getRoot();
-
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_complete_purchase, container, false);
     }
 }
