+++ app/src/main/java/com/example/apparty/DetailEventItemFragment.java
@@ -1,4 +1,4 @@
-package com.example.apparty.fragments;
+package com.example.apparty;
 
 import android.os.Bundle;
 
@@ -8,25 +8,25 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.apparty.R;
+import com.example.apparty.databinding.FragmentDetailEventItemBinding;
 
-public class DetailPurchaseFragment extends Fragment {
+public class DetailEventItemFragment extends Fragment {
 
-    public DetailPurchaseFragment() {
-        // Required empty public constructor
-    }
+    private FragmentDetailEventItemBinding binding;
+
+    public DetailEventItemFragment() {}
 
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_detail_purchase, container, false);
+        binding = FragmentDetailEventItemBinding.inflate(inflater, container, false);
+        return binding.getRoot();
+
     }
 }
