+++ app/src/main/java/com/example/apparty/DetailPurchaseFragment.java
@@ -1,4 +1,4 @@
-package com.example.apparty.fragments;
+package com.example.apparty;
 
 import android.os.Bundle;
 
@@ -8,23 +8,23 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.apparty.R;
+public class DetailPurchaseFragment extends Fragment {
 
-public class CompletePurchaseFragment extends Fragment {
-
-    public CompletePurchaseFragment() {
+    public DetailPurchaseFragment() {
         // Required empty public constructor
     }
+
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_complete_purchase, container, false);
+        return inflater.inflate(R.layout.fragment_detail_purchase, container, false);
     }
 }
