+++ app/src/main/java/com/example/apparty/LoginFragment.java
@@ -1,4 +1,4 @@
-package com.example.apparty.fragments;
+package com.example.apparty;
 
 import android.os.Bundle;
 
@@ -6,11 +6,13 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.fragment.NavHostFragment;
 
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.example.apparty.R;
+import com.example.apparty.databinding.FragmentLoginBinding;
+import com.example.apparty.databinding.FragmentSearchEventsBinding;
 
 
 public class LoginFragment extends Fragment {
