+++ app/src/main/java/com/example/apparty/PurchaseFragment.java
@@ -1,4 +1,4 @@
-package com.example.apparty.fragments;
+package com.example.apparty;
 
 import android.os.Bundle;
