+++ app/src/main/java/com/example/apparty/model/Address.java
@@ -1,18 +1,14 @@
 package com.example.apparty.model;
 
-import androidx.room.Entity;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Entity (tableName = "address")
 public class Address {
     private int id;
     private String address;
