+++ app/src/main/java/com/example/apparty/model/Filter.java
@@ -1,7 +1,5 @@
 package com.example.apparty.model;
 
-import androidx.room.Entity;
-
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Date;
@@ -13,7 +11,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
-@Entity
 @Getter
 @Setter
 @NoArgsConstructor
