+++ app/src/main/java/com/example/apparty/model/Stock.java
@@ -1,8 +1,5 @@
 package com.example.apparty.model;
 
-import androidx.room.Entity;
-import androidx.room.PrimaryKey;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -12,14 +9,11 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Entity (tableName = "stock")
 public class Stock {
-    @PrimaryKey(autoGenerate = true)
     private int id;
     private String type;
     private int totalQuantity;
     private int availableQuantity;
     private double price;
-    private int idEvent;
 
 }
