+++ app/src/main/java/com/example/apparty/model/Ticket.java
@@ -1,13 +1,10 @@
 package com.example.apparty.model;
 
-import androidx.room.Entity;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-@Entity
 @Getter
 @Setter
 @NoArgsConstructor
