+++ app/src/main/java/com/example/apparty/model/User.java
@@ -1,10 +1,5 @@
 package com.example.apparty.model;
 
-import androidx.room.Entity;
-import androidx.room.PrimaryKey;
-
-import org.jetbrains.annotations.PropertyKey;
-
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,9 +9,7 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-@Entity (tableName = "user")
 public class User {
-    @PrimaryKey(autoGenerate = true)
     private int id;
     private String name;
     private String surname;
