+++ app/src/main/java/com/example/apparty/model/Utils.java
@@ -1,6 +1,5 @@
 package com.example.apparty.model;
 
-
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
