--- app/src/main/java/com/example/apparty/model/relations/EventStock.java
@@ -1,17 +1,0 @@
-package com.example.apparty.model.relations;
-
-
-import androidx.room.Embedded;
-import androidx.room.Relation;
-
-import com.example.apparty.model.Event;
-import com.example.apparty.model.Stock;
-
-import java.util.List;
-
-public class EventStock {
-    @Embedded public Event event;
-    @Relation(parentColumn = "id_event", entityColumn = "idEvent", entity = Stock.class)
-    List<Stock> stockList;
-
-}
