+++ app/src/main/res/layout/fragment_detail_event_item.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".fragments.DetailEventItemFragment">
+    tools:context=".DetailEventItemFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
