+++ app/src/main/res/layout/fragment_detail_purchase.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/white"
     android:orientation="vertical"
-    tools:context=".fragments.DetailPurchaseFragment">
+    tools:context=".DetailPurchaseFragment">
 
     <TextView
         android:layout_width="match_parent"
