+++ app/src/main/res/layout/fragment_event_detail.xml
@@ -7,7 +7,7 @@
     android:background="@color/dark_blue"
     android:orientation="vertical"
     android:padding="10dp"
-    tools:context=".fragments.EventDetailFragment">
+    tools:context=".EventDetailFragment">
 
     <com.google.android.material.card.MaterialCardView
         android:layout_width="match_parent"
