+++ app/src/main/res/layout/fragment_event_results.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.EventResultsFragment">
+    tools:context=".EventResultsFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -21,8 +21,7 @@
             android:textAppearance="?attr/textAppearanceTitleMedium"
             android:textSize="20sp"
             android:layout_marginTop="15dp"
-            android:layout_marginBottom="5dp"
-            android:layout_marginHorizontal="10dp"/>
+            android:layout_marginBottom="5dp"/>
         <TextView
             android:id="@+id/textResume"
             android:layout_width="match_parent"
@@ -33,7 +32,6 @@
             android:layout_marginBottom="8dp"
             android:textColor="@color/white"
             android:textAppearance="?attr/textAppearanceBody2"
-            android:layout_marginHorizontal="10dp"
             />
 
         <com.google.android.gms.maps.MapView
@@ -44,7 +42,8 @@
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerResult"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"/>
+            android:layout_height="wrap_content"
+            android:layout_margin="10dp"/>
 
     </LinearLayout>
