+++ app/src/main/res/layout/fragment_filter_dialog.xml
@@ -7,7 +7,7 @@
     android:background="@color/white"
     android:orientation="vertical"
     android:padding="5dp"
-    tools:context=".fragments.FilterDialogFragment">
+    tools:context=".FilterDialogFragment">
 
     <TextView
         android:layout_width="match_parent"
