+++ app/src/main/res/layout/fragment_login.xml
@@ -6,7 +6,7 @@
     android:background="@color/dark_blue"
     android:orientation="vertical"
     android:paddingHorizontal="30dp"
-    tools:context=".fragments.LoginFragment">
+    tools:context=".LoginFragment">
 
     <TextView
         android:id="@+id/title"
