+++ app/src/main/res/layout/fragment_purchase.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".fragments.PurchaseFragment">
+    tools:context=".PurchaseFragment">
 
     <com.google.android.material.tabs.TabLayout
         android:id="@+id/tablayout"
