+++ app/src/main/res/layout/fragment_registry.xml
@@ -6,7 +6,7 @@
     android:background="@color/dark_blue"
     android:orientation="vertical"
     android:paddingHorizontal="30dp"
-    tools:context=".fragments.RegistryFragment">
+    tools:context=".RegistryFragment">
 
     <TextView
         android:id="@+id/title"
