+++ app/src/main/res/layout/fragment_search_events.xml
@@ -7,7 +7,7 @@
     android:background="@color/dark_blue"
     android:orientation="vertical"
     android:paddingHorizontal="15sp"
-    tools:context=".fragments.SearchEventsFragment">
+    tools:context=".SearchEventsFragment">
 
     <com.google.android.material.card.MaterialCardView
         android:id="@+id/card"
