+++ app/src/main/res/navigation/login_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.example.apparty.fragments.LoginFragment"
+        android:name="com.example.apparty.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login" >
         <action
@@ -20,7 +20,7 @@
 
     <fragment
         android:id="@+id/registryFragment"
-        android:name="com.example.apparty.fragments.RegistryFragment"
+        android:name="com.example.apparty.RegistryFragment"
         android:label="fragment_registry"
         tools:layout="@layout/fragment_registry" >
         <action
