+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/searchEvents"
-        android:name="com.example.apparty.fragments.SearchEventsFragment"
+        android:name="com.example.apparty.SearchEventsFragment"
         android:label="fragment_search_events"
         tools:layout="@layout/fragment_search_events" >
         <action
@@ -23,7 +23,7 @@
 
     <fragment
         android:id="@+id/eventResultsFragment"
-        android:name="com.example.apparty.fragments.EventResultsFragment"
+        android:name="com.example.apparty.EventResultsFragment"
         android:label="fragment_event_results"
         tools:layout="@layout/fragment_event_results">
         <action
@@ -33,7 +33,7 @@
 
     <fragment
         android:id="@+id/eventDetailFragment"
-        android:name="com.example.apparty.fragments.EventDetailFragment"
+        android:name="com.example.apparty.EventDetailFragment"
         android:label="fragment_event_detail"
         tools:layout="@layout/fragment_event_detail">
         <action
@@ -42,7 +42,7 @@
     </fragment>
     <fragment
         android:id="@+id/purchaseFragment"
-        android:name="com.example.apparty.fragments.PurchaseFragment"
+        android:name="com.example.apparty.PurchaseFragment"
         android:label="fragment_purchase"
         tools:layout="@layout/fragment_purchase" />
